lit: change read and parse methods to top level functions
We leave ParseVal and ParseMut as methods of Reg, so the
distinction to the usual Parse functions is more obvious.

We need the Reg parameter to parse expressions too. So we ended up
with exp.ParseX(*lit.Reg, …) functions and (*lit.Reg).ParseX(…)
methods, because we can obviously not use methods in exp.

The ides to make lit parse methods was to make it clear that reg is
strictly required. However the difference in style alone is enough
as a reason for the change.

The lit registry is also now obviously a core concept of xelf that
any user of the needs to understand anyway.