typ: change how tupl type body is used
We use the tupl type to represent sequences of a certain type of
expressions. Often enough this a simple unnamed element type tupl?|tag
that represents of optional sequence of tags. But sometimes we want to
represent a sequence of multiple typed elements. One such case is the if
statement where we want to match a non-zero sequence of condition and
action expressions:

	<form if <tupl cond:any act:exp|@1> else?:exp|@1 @1>

Before we used a param type body for both variants, we now change it so we
have the tupl type use both the ElBody and ParamBody depending on the
context.

We already tried to discern both context with two different badly-named
helpers TuplList and TuplRec - now renamed to ElemTupl and ParamTupl.