typ: serialise a type as type literal in xelf context
This allows us to write and parse type literal when encoded using xelf
notation. This is only reasonable and a huge help not only for testing.

Types are still written as string literal when encoded as json. We could
map it to a json object, then again as string it is concise and clearly
noticeable as type literal. The code to parse a type is also not too
hard to replicate and probably already available if it matters.