lib: change mut spec to allow edit paths on any value and add append spec
We have cool new edit paths and list ops used for literal deltas. We want
to make them available to the standard libraries mut spec. This works just
fine for keyrs. All keyr values are also mutable values and we already
allow key paths. Edit paths work for all values, and any keyr field value
is assignable, therefor we should also change the mut spec to accept any
type of value. Then the most common mutation assignment should be
expressible as (mut .a 1) and not only as the edit path (mut .a .:1). All
other mutations use edit paths (mut prod name:"Ketchup").

This new assignment syntax must be consistent for all values, but clashes
with the append style syntax that mut supports to date. We want append in
some form, mainly used with dyn sugar to be able to write ([] 1 2 3). The
only clean solution is to introduce an append spec, like in previous
designs had. The plain tupl is now only used by append and will be change
to a single optional argument for the simple assignment variant. mut can
still be used to append (mut [] .+:[1 2 3])

When both assign and edits are given e.g: (mut {} .defaults addr:'…') the
assign value is first assigned to the target, then the updates are
applied. This has some valid use cases like initializing with values
before applying edits.

Append defies the three letter naming scheme because apd is not easily
recognizable as append. Another reason is that we intend to use it as dyn
sugar without its name.