lit: introduce the delta list type and rename the operation to diff
Also actually implement edit path in the mut spec.

I thought for a time that using maps had more usability upsides than
downsides, because we would need to design most things around dicts
with the assumption that we may have a random key order. Apart from the
issue of testing unordered dicts, when writing and debugging code in xelf
it is generally cognitively easier to have the output order match the
input order. Specifically for deltas we may work dicts that are later
resolved to be strcs.