exp: api breaking! add context to prog and eval helpers
We add shorter aliases in the exp package.

We do not check for timeout in the program code for now.
But implementations can easily check the context in custom specs.

For example the daql qry package can pass on the context for database
queries and check in between multiple queries.