exp: do not embed lit val in exp lit and add tags select to avoid exp dicts
Previously we embedded val so we could reuse the path select on exp lit
inside dicts for the func env in a bad attempt to preserve more type
information. This was confusing and we ideally want to avoid confusion.

We introduced tags with methods to select a path from a list of arguments
and preserving the resolved type and changed the embedded val to a named
field.