exp: similar to raw types, spec literals must have the varspec type
More specifically spec literals must have an instantiated type or varspec.
So a safe convention is always to use varspec and vartyp for external
literals.

When resolving a spec symbol, the program instantiates the signature
and stores it in the lit type, that is then reused in call resolution.
I am not completely confident the reuse is ok, but attempts to find error
conditions so far.