exp: similar to raw types, spec literals must have a plain spec type
More specifically spec literals must have an instantiated type or spec.
A safe convention is to always use spec and vartyp for external literals.

When resolving a spec symbol, the program instantiates the signature
and stores it in the lit type, that is then reused in call resolution.
I am not completely confident the reuse is ok, but attempts to find error
conditions so far.